Back/requirements.txt
fastapi==0.95.2
uvicorn==0.22.0
python-jose==3.3.0
pydantic==1.10.7
numpy==1.24.3
opencv-python==4.7.0.72
mediapipe==0.10.0
pyttsx3==2.90
scikit-learn==1.2.2
python-multipart==0.0.6